requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
numpy
tqdm
sentence-transformers
faiss-cpu
pypdf
openai
requests
